python/requirements.txt
# Core
pandas>=2.1
numpy>=1.26
scipy>=1.11
matplotlib>=3.8
seaborn>=0.13

# Modeling
statsmodels>=0.14
lifelines>=0.28
scikit-learn>=1.3

# Config & reporting
PyYAML>=6.0
Jinja2>=3.1
python-docx>=1.1

# Jupyter (optional)
jupyterlab>=4.0
ipykernel>=6.29
